requirements.txt
ezdxf>=1.1.4
numpy>=1.23
scipy>=1.10
networkx>=3.2
pyyaml>=6.0